Audit workspace lifecycle provider boundaries
